fix: no hard coded url for openfoodfacts.org also avoid circular dependency between api and signals
