+++ web-components/src/api/folksonomy.ts
@@ -11,7 +11,7 @@ import {
   PropertyDeleteRequest,
   PropertyClashCheck,
 } from "../types/folksonomy"
-import { folksonomyConfiguration } from "../signals/folksonomy"
+import { folksonomyConfiguration, userInfo, userInfoLoading } from "../signals/folksonomy"
 
 // Constants for localStorage
 const FOLKSONOMY_BEARER_TOKEN_KEY = "folksonomy-bearer-token"
@@ -314,6 +314,45 @@ async function getUserInfo(): Promise<UserInfo> {
   }
 }
 
+/**
+ * Fetch user info and update the shared signal
+ * Only fetches if not already loading or if force is true
+ */
+async function fetchUserInfo(force: boolean = false): Promise<UserInfo | null> {
+  // Don't fetch if already loading (unless forced)
+  if (userInfoLoading.get() && !force) {
+    return userInfo.get()
+  }
+
+  // Don't fetch if we already have user info (unless forced)
+  if (userInfo.get() && !force) {
+    return userInfo.get()
+  }
+
+  userInfoLoading.set(true)
+
+  try {
+    const fetchedUserInfo = await getUserInfo()
+    userInfo.set(fetchedUserInfo)
+    return fetchedUserInfo
+  } catch (error) {
+    console.error("Error fetching user info:", error)
+    // User might not be authenticated, which is fine
+    userInfo.set(null)
+    return null
+  } finally {
+    userInfoLoading.set(false)
+  }
+}
+
+/**
+ * Clear user info (useful for logout)
+ */
+export function clearUserInfo() {
+  userInfo.set(null)
+  userInfoLoading.set(false)
+}
+
 async function replaceValue(request: ValueRenameRequest): Promise<{ success: boolean }> {
   try {
     const response = await makeAuthenticatedRequest(getApiUrl("/admin/value/replace"), {
@@ -430,6 +469,7 @@ export default {
   fetchValues,
   authByCookie,
   getUserInfo,
+  fetchUserInfo,
   replaceValue,
   deleteValue,
   checkPropertyClash,
